Tweak the imported test rule code to work in glean-core
