+++ glean-core/android/src/main/java/mozilla/telemetry/glean/testing/GleanTestRule.kt
@@ -2,12 +2,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.service.glean.testing
+package mozilla.telemetry.glean.testing
 
 import android.content.Context
 import androidx.work.testing.WorkManagerTestInitHelper
-import mozilla.components.service.glean.Glean
-import mozilla.components.service.glean.config.Configuration
+import mozilla.telemetry.glean.Glean
+import mozilla.telemetry.glean.config.Configuration
 import org.junit.rules.TestWatcher
 import org.junit.runner.Description
